Add format-specifier warnings to SkDebugf.
This CL fixes up many existing format-specifier violations.
Note that GCC has a warning for formatting nothing, so existing calls to
`SkDebugf("")` have been removed, or replaced with `SkDebugf("%s", "")`.
These were apparently meant to be used as a place to set a breakpoint.

+++ modules/skplaintexteditor/app/editor_application.cpp
@@ -209,7 +209,7 @@ struct EditorLayer : public sk_app::Window::Layer {
             switch (c) {
                 case 'p':
                     for (StringView str : fEditor.text()) {
-                        SkDebugf(">>  '%.*s'\n", str.size, str.data);
+                        SkDebugf(">>  '%.*s'\n", (int)str.size, str.data);
                     }
                     return true;
                 case 's':
